updated api tests
